Provide absolute and relative timestamps
To maintain backwards compatibility, leave the `timestamp` key alone as
`timestamp_absolute` and present a `timestamp_relative`.

If we wanted, we could just agree on a single global absolute starting
timestamp, pass around only relative timestamps, and determine absolute
timestamps as differences. However, this might not be a good idea due
to stream instability and dropping, so providing absolute and relative
times separately might be worth the negligible additional data juggling.